requirements.txt
requests
beautifulsoup4
openai
google-generativeai
python-dotenv
